backend/requirements.txt
fastapi==0.128.4
uvicorn[standard]==0.40.0
python-multipart==0.0.22
pydantic-settings==2.12.0
pymupdf==1.26.7
python-docx==1.2.0
openai==2.17.0
numpy==2.4.2
